Image in Details now load from api
