+++ app/src/main/java/com/example/learnabird/DetailActivity.java
@@ -9,6 +9,8 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.example.learnabird.AsyncTasks.AsyncLoadImage;
+
 public class DetailActivity extends AppCompatActivity {
 
     private ImageView iv_imgPreview;
@@ -16,6 +18,7 @@ public class DetailActivity extends AppCompatActivity {
     private TextView tv_birdDetails;
     private Button btn_playSound;
     private MediaPlayer mediaPlayer;
+    private int birdId;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -30,11 +33,13 @@ public class DetailActivity extends AppCompatActivity {
 
         final Bundle bundle = getIntent().getExtras();
         if (bundle != null) {
+            birdId = bundle.getInt("birdId");
             tv_birdName.setText(bundle.getString("birdName"));
-            iv_imgPreview.setImageResource(bundle.getInt("birdPic"));
+
+            new AsyncLoadImage(iv_imgPreview).execute(bundle.getString("birdPic"));
             tv_birdDetails.setText(bundle.getString("birdDetails"));
             getSupportActionBar().setTitle("Learn A Bird : " + bundle.getString("birdName"));
-            mediaPlayer = (MediaPlayer) MediaPlayer.create(this, bundle.getInt("birdSounds"));
+            mediaPlayer = (MediaPlayer) MediaPlayer.create(this, bundle.getInt("birdSound"));
         }
 
         btn_playSound.setOnClickListener(new View.OnClickListener() {
