+++ app/src/main/java/com/example/learnabird/ListAdapter.java
@@ -1,9 +1,6 @@
 package com.example.learnabird;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.os.AsyncTask;
 import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,24 +9,19 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
+import com.example.learnabird.AsyncTasks.AsyncLoadImage;
 
 public class ListAdapter extends ArrayAdapter {
 
     private String[] birdNames;
     private String[] birdPics;
     private Context context;
-    private String hostUrl;
 
-    public ListAdapter(Context context, String[] names, String[] pics, String hostUrl) {
+    public ListAdapter(Context context, String[] names, String[] pics) {
         super(context, R.layout.list_view_item);
         this.birdNames = names;
         this.birdPics = pics;
         this.context = context;
-        this.hostUrl = hostUrl;
     }
 
     @Override
@@ -51,7 +43,7 @@ public class ListAdapter extends ArrayAdapter {
         else {
             viewHolder = (ViewHolder)convertView.getTag();
         }
-        new LoadImage(viewHolder.mBird).execute(birdPics[position]);
+        new AsyncLoadImage(viewHolder.mBird).execute(birdPics[position]);
         viewHolder.mName.setText(birdNames[position]);
         return convertView;
     }
@@ -59,37 +51,6 @@ public class ListAdapter extends ArrayAdapter {
     static class ViewHolder{
         ImageView mBird;
         TextView mName;
-    }
-
-    public class LoadImage extends AsyncTask<String, Void, Bitmap> {
-
-        ImageView target;
-
-        public LoadImage(ImageView imageView){
-            this.target = imageView;
-        }
-
-        @Override
-        protected Bitmap doInBackground(String... images) {
-
-            Bitmap result = null;
-            try {
-                URL url = new URL(hostUrl+images[0]);
-                InputStream in = url.openStream();
-                result = BitmapFactory.decodeStream(in);
-            } catch (MalformedURLException e) {
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            return result;
-        }
-
-        @Override
-        protected void onPostExecute(Bitmap bitmap) {
-            super.onPostExecute(bitmap);
-            target.setImageBitmap(bitmap);
-        }
     }
 
 }
